Initial sketch of what I'm thinking
I put a bunch of time into the readme to go over my thoughts about
what the library is trying to provide, and have a small amount of
rough code that might just work (very much WIP, though).

There are some decisions to be made as far as the interface and
how the macros behave, but the skeleton is there.