Working code!
This commit has the initial skeleton of the header file and a basic
example that demonstrates nested async functions and even "parallel"
fork-join usage with &.

I went with the "everything on the stack" concept, and using a local
struct to unpack the data turned out to be a lot smoother than I
anticipated.

There are two definitions for async_begin() to let you disable the stack
bounds checking for well-tested code (saves runtime overhead and some
memory overhead, so it is a worthwhile addition).  I have not tested
that implementation yet, so that's a TODO.

I tried very hard to find a way to avoid struct dereferencing for the
locals, but nothing made sense in terms of overhead and complexity (and
safety).  Packing them all into a struct and dereferencing that makes
the bounds checking and un-packing easier (and likely more performant).
It is not possible to define new preprocessor values in a macro, so my
initial thought of faking locals is not possible :(

To help with the look I made a _() macro to dereference the local value
for you, but that just looks slightly better at the cost of typing ease.
I'm not sure that this is a deal breaker though.  Locals are more of a
"when you absolutely need them" thing, not a "use it just like you would
ordinary code" thing.  You can't win them all.

I created a helper macro for doing a hex dump of the stacks because I am
dumb and can't count (and can't remember to pop things off of a stack,
apparently).  I blame excitement for pushing me into testing before I
actually thought about everything ;)

Oh, and this code definitely lives up to the macro magic hype.  I had to
use a FOR_EACH() macro to implement the syntax for locals that I wanted.
It isn't really that nasty, and the one edge case where it doesn't
exactly work completely correctly (empty lists) does not exist in this
library, because I always have the "spot" value on the stack to keep
track of the async state/location in the function.

None of the other helper features or runtime functionality is
implemented, just the core async stuff with my stack context additions.
I added some more ideas/thoughts to the readme.

It's still very rough around the edges, but it works, so now we have
somewhere to start.  Or I could go off into the weeds on a new concept
trying to implement futures and promises in C in an embedded-friendly
way.  Who knows.